+++ tests/utilities/test_auto_restart.py
@@ -35,6 +35,7 @@ from torch.utils.data.dataset import Dataset, IterableDataset
 
 import tests.helpers.utils as tutils
 from pytorch_lightning import Callback, LightningModule, seed_everything, Trainer
+from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.utilities.auto_restart import (
     _add_capture_metadata_collate,
     _dataloader_load_state_dict,
@@ -45,7 +46,7 @@ from pytorch_lightning.utilities.auto_restart import (
     FastForwardSampler,
     MergedIteratorState,
 )
-from pytorch_lightning.utilities.enums import AutoRestartBatchKeys
+from pytorch_lightning.utilities.enums import _FaultTolerantMode, AutoRestartBatchKeys
 from pytorch_lightning.utilities.exceptions import ExitGracefullyException, MisconfigurationException
 from pytorch_lightning.utilities.fetching import DataFetcher
 from pytorch_lightning.utilities.imports import _fault_tolerant_training
@@ -1195,7 +1196,7 @@ def test_auto_restart_under_signal(on_last_batch, val_check_interval, failure_on
         assert "dataloader_state_dict" in state_dict
 
 
-def test__is_obj_stateful():
+def test_is_obj_stateful():
     class StatefulClass:
         def state_dict(self):
             pass
@@ -1216,9 +1217,29 @@ def test__is_obj_stateful():
     obj = NotStatefulClass()
     assert not _is_obj_stateful(obj)
 
-    class NotStatefulClass:
-        def load_state_dict(self):
+    class NotStateful2Class:
+        def load_state_dict(self, state_dict):
             pass
 
-    obj = NotStatefulClass()
+    obj = NotStateful2Class()
     assert not _is_obj_stateful(obj)
+
+
+def test_fault_tolerant_mode_enum():
+    with mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "0"}):
+        assert _FaultTolerantMode.DISABLED == _FaultTolerantMode.detect_current_mode()
+        assert not TrainerState()._fault_tolerant_mode.is_enabled
+
+    with mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "1"}):
+        assert _FaultTolerantMode.AUTOMATIC == _FaultTolerantMode.detect_current_mode()
+        assert TrainerState()._fault_tolerant_mode.is_automatic
+
+    with mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "MANUAL"}):
+        assert _FaultTolerantMode.MANUAL == _FaultTolerantMode.detect_current_mode()
+        assert TrainerState()._fault_tolerant_mode.is_manual
+
+    with pytest.raises(
+        MisconfigurationException, match="The environment flag `PL_FAULT_TOLERANT_TRAINING` should be either"
+    ):
+        with mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "3"}):
+            _FaultTolerantMode.detect_current_mode()